+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # CHANGELOG.md
 
-## v0.38.1
+## v0.39.0
  - Ability to execute sql for URL paths with another extension. If you create sitemap.xml.sql, it will be executed for example.com/sitemap.xml
  - Display source line info in errors even when the database does not return a precise error position. In this case, the entire problematic SQL statement is referenced.
  - The shell with a vertical sidebar can now have "active" elements, just like the horizontal header bar.
+ - New `edit_url`, `delete_url`, and `custom_actions` properties in the [table](https://sql-page.com/component.sql?component=table) component to easily add nice icon buttons to a table.
 
 ## v0.38.0
  - Added support for the Open Database Connectivity (ODBC) standard.